+++ src/main/java/webserver/router/ResponseHandlerRouter.java
@@ -1,4 +1,4 @@
-package webserver.parser;
+package webserver.router;
 
 import config.Config;
 import webserver.handler.response.*;
@@ -7,12 +7,12 @@ import webserver.request.*;
 import java.io.DataOutputStream;
 import java.io.OutputStream;
 
-public class ResponseHandlerParser {
+public class ResponseHandlerRouter {
     private DataOutputStream responseWriter;
     private Config config;
     private HttpRequest httpRequest;
 
-    public ResponseHandlerParser(OutputStream out, Config config, HttpRequest httpRequest) {
+    public ResponseHandlerRouter(OutputStream out, Config config, HttpRequest httpRequest) {
         this.responseWriter = new DataOutputStream(out);
         this.config = config;
         this.httpRequest = httpRequest;
